+++ src/com/sundays/chat/io/jdbc/ChannelDataUpdater.java
@@ -65,14 +65,14 @@ public class ChannelDataUpdater {
 		return permissionArray;		
 	}
 	
-	public static Map<Byte, String> decompressRankNamesV1 (byte[] rankNames) {
-		Map<Byte, String> nameArray = new HashMap<>();
+	public static Map<Integer, String> decompressRankNamesV1 (byte[] rankNames) {
+		Map<Integer, String> nameArray = new HashMap<>();
 		try {
 			ByteArrayExtractor extractor = new ByteArrayExtractor(rankNames);//Converts the names list into a byte array
             if (extractor.getByte() != 12) {//The first value in v1 compressions was the number of ranks. This used to be fixed at 12.
             	throw new IllegalArgumentException("Invalid rank names data submitted to ChannelDataUpdate.decompressRankNames()");
             }
-            byte i=0;
+            int i=0;
             while (true) {
             	String name;
                 try {
@@ -90,8 +90,8 @@ public class ChannelDataUpdater {
 		return nameArray;
 	}
 	
-	public static Map<Byte, String> decompressRankNamesV2 (byte[] rankNames) {
-		Map<Byte, String> nameArray = new HashMap<>();
+	public static Map<Integer, String> decompressRankNamesV2 (byte[] rankNames) {
+		Map<Integer, String> nameArray = new HashMap<>();
 		try {
 			ByteArrayExtractor extractor = new ByteArrayExtractor(rankNames);//Converts the names list into a byte array
             if (extractor.getShort() != 2) {//Reads the rank name archive version
@@ -100,7 +100,7 @@ public class ChannelDataUpdater {
             }
             while (true) {
                 String name;
-                byte rank;
+                int rank;
             	try {
             		rank = extractor.getByte();
                     name = extractor.getUTFString();
@@ -136,11 +136,11 @@ public class ChannelDataUpdater {
 		return bldr.getByteArray();
 	}
 	
-	private byte[] compressRankNamesV2 (Map<Byte, String> rankNames) throws IOException {
+	private byte[] compressRankNamesV2 (Map<Integer, String> rankNames) throws IOException {
 		ByteArrayBuilder bldr = new ByteArrayBuilder();
 		bldr.writeShort(Settings.RANK_NAME_VERSION);//Rank names version
-		for (Entry<Byte, String> name : rankNames.entrySet()) {
-			bldr.writeByte(name.getKey());
+		for (Entry<Integer, String> name : rankNames.entrySet()) {
+			bldr.writeByte(name.getKey().byteValue());
         	bldr.writeUTFString(name.getValue());
         }
 		return bldr.getByteArray();
